+++ test/jest/__mock__/finalFormComponents.mock.js
@@ -1,7 +1,13 @@
 import React from 'react';
 
 jest.mock('react-final-form', () => ({
-  Field: jest.fn(({ label, component, children, ...rest }) => (
+  Field: jest.fn(({
+    label,
+    component,
+    children,
+    dataOptions,
+    ...rest
+  }) => (
     <div {...rest}>
       {label}
       {component()}