+++ test/jest/__mock__/stripesSmartComponents.mock.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 jest.mock('@folio/stripes/smart-components', () => ({
+  ConfigManager: jest.fn(() => null),
   ViewMetaData: jest.fn(({ metadata, ...rest }) => (
     <div {...rest}>{metadata.createdDate}</div>
   )),